requirements.txt
torch
torchvision
numpy
scikit-learn
matplotlib
tqdm
pillow
